InKey Backend/requirements.txt
bashlex==0.14
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudinary==1.21.0
compiledb==0.10.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
deprecation==2.1.0
Django==2.2.6
django-annoying==0.10.6
django-cloudinary-storage==0.2.3
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt
drf-yasg==1.17.1
eyed3==0.9.7
filetype==1.2.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.17.4
packaging==20.3
Pillow==7.1.1
pydot==1.4.1
PyJWT==1.7.1
pyparsing==2.4.5
pytz==2019.3
PyYAML==5.2
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.0
shutilwhich==1.1.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.8
uuid==1.30
whitenoise==5.0.1